changes for requeriments of corebeat
