+++ bethink/src/hooks/useAnimationLike.js
@@ -4,7 +4,6 @@ import {Animated} from 'react-native';
 export const useAnimationLike = setShowAnimation => {
   const move = useRef(new Animated.Value(250)).current;
   const opacity = useRef(new Animated.Value(1)).current;
-
   const animation = () => {
     Animated.timing(move, {
       toValue: 50,
