+++ bethink/src/screens/home/HomeScreen.js
@@ -6,18 +6,14 @@ import {getAllNotes} from '../../actions/notes';
 import {ContainerMain} from '../../components/home/ContainerMain';
 import {Date} from '../../components/home/Date';
 import {DisplayUser} from '../../components/home/DisplayUser';
-import {LoadingNotes} from '../../components/home/LoadingNotes';
-import {LoveNotes} from '../../components/home/LoveNotes';
-import {NoLoves} from '../../components/home/NoLoves';
 import {Reminders} from '../../components/home/Reminders';
 import {Logo} from '../../components/home/Logo';
+import {ConditionNotes} from '../../components/home/ConditionNotes'
 
 export const HomeScreen = () => {
   const {notes, loadingNotes} = useSelector(state => state.notes);
-  const [refreshing, setRefreshing] = useState(false);
   const {uid} = useSelector(state => state.auth);
   const dispatch = useDispatch();
-
   const notesLoveIt = notes.filter(note => note.love === true);
 
   useFocusEffect(
@@ -26,19 +22,9 @@ export const HomeScreen = () => {
     }, []),
   );
 
-  const onRefresh = () => {
-    setRefreshing(true);
-    dispatch(getAllNotes(uid));
-    setRefreshing(false);
-  };
-
   return (
     <ContainerMain>
-      <ScrollView
-        showsVerticalScrollIndicator={false}
-        refreshControl={
-          <RefreshControl refreshing={refreshing} onRefresh={onRefresh} />
-        }>
+      <ScrollView showsVerticalScrollIndicator={false}>
         <View style={styles.container}>
           <Logo />
           <DisplayUser />
@@ -46,13 +32,7 @@ export const HomeScreen = () => {
             <Date />
           </View>
           <Reminders />
-          {loadingNotes ? (
-            <LoadingNotes />
-          ) : notesLoveIt.length > 0 ? (
-            <LoveNotes notes={notesLoveIt} />
-          ) : (
-            <NoLoves />
-          )}
+          <ConditionNotes notesLoveIt={notesLoveIt} loadingNotes={loadingNotes} />
         </View>
       </ScrollView>
     </ContainerMain>
